# Use parallel pytest in the CI checks