--- a/check/pytest-and-incremental-coverage
+++ b/check/pytest-and-incremental-coverage
@@ -28,7 +28,14 @@ cd "$(git rev-parse --show-toplevel)" || exit 1
 
 PYTEST_ARGS=()
 ANALYZE_COV=1
-for arg in "${@:2}"; do
+BASEREV=""
+
+if [[ "$1" == [^-]* ]]; then
+    BASEREV="$1"
+    shift
+fi
+
+for arg in "$@"; do
     if [[ "${arg}" == "--no-analyze" ]]; then
         ANALYZE_COV=0
     else
@@ -37,12 +44,12 @@ for arg in "${@:2}"; do
 done
 
 # Figure out which revision to compare against.
-if [ ! -z "$1" ] && [[ $1 != -* ]]; then
-    if [ "$(git cat-file -t $1 2> /dev/null)" != "commit" ]; then
-        echo -e "\033[31mNo revision '$1'.\033[0m" >&2
+if [ -n "${BASEREV}" ]; then
+    if [ "$(git cat-file -t "${BASEREV}" 2> /dev/null)" != "commit" ]; then
+        echo -e "\033[31mNo revision '${BASEREV}'.\033[0m" >&2
         exit 1
     fi
-    rev=$1
+    rev="${BASEREV}"
 elif [ "$(git cat-file -t upstream/master 2> /dev/null)" == "commit" ]; then
     rev=upstream/master
 elif [ "$(git cat-file -t origin/master 2> /dev/null)" == "commit" ]; then