--- a/dev_tools/bash_scripts_test.py
+++ b/dev_tools/bash_scripts_test.py
@@ -353,6 +353,9 @@ def test_pytest_and_incremental_coverage_branch_selection(tmpdir_factory):
         'INTERCEPTED check/pytest '
         '--actually-quiet --rigetti-integration --cov '
         '--cov-config=dev_tools/conf/.coveragerc\n'
+        'The annotate command will be removed in a future version.\n'
+        'Get in touch if you still use it: ned@nedbatchelder.com\n'
+        'No data to report.\n'
         'INTERCEPTED '
         'python dev_tools/check_incremental_coverage_annotations.py HEAD\n'
     )
@@ -375,10 +378,13 @@ def test_pytest_and_incremental_coverage_branch_selection(tmpdir_factory):
     assert result.exit_code == 0
     assert result.out == (
         'INTERCEPTED check/pytest '
-        '--actually-quiet --rigetti-integration --cov --cov-report=annotate '
+        '--actually-quiet --rigetti-integration --cov '
         '--cov-config=dev_tools/conf/.coveragerc\n'
-        'INTERCEPTED python '
-        'dev_tools/check_incremental_coverage_annotations.py master\n'
+        'The annotate command will be removed in a future version.\n'
+        'Get in touch if you still use it: ned@nedbatchelder.com\n'
+        'No data to report.\n'
+        'INTERCEPTED '
+        'python dev_tools/check_incremental_coverage_annotations.py master\n'
     )
     assert result.err == "Comparing against revision 'master'.\n"
 
@@ -391,10 +397,13 @@ def test_pytest_and_incremental_coverage_branch_selection(tmpdir_factory):
     assert result.exit_code == 0
     assert result.out == (
         'INTERCEPTED check/pytest '
-        '--actually-quiet --rigetti-integration --cov --cov-report=annotate '
+        '--actually-quiet --rigetti-integration --cov '
         '--cov-config=dev_tools/conf/.coveragerc\n'
-        'INTERCEPTED python '
-        'dev_tools/check_incremental_coverage_annotations.py origin/master\n'
+        'The annotate command will be removed in a future version.\n'
+        'Get in touch if you still use it: ned@nedbatchelder.com\n'
+        'No data to report.\n'
+        'INTERCEPTED '
+        'python dev_tools/check_incremental_coverage_annotations.py origin/master\n'
     )
     assert result.err == "Comparing against revision 'origin/master'.\n"
 
@@ -407,10 +416,13 @@ def test_pytest_and_incremental_coverage_branch_selection(tmpdir_factory):
     assert result.exit_code == 0
     assert result.out == (
         'INTERCEPTED check/pytest '
-        '--actually-quiet --rigetti-integration --cov --cov-report=annotate '
+        '--actually-quiet --rigetti-integration --cov '
         '--cov-config=dev_tools/conf/.coveragerc\n'
-        'INTERCEPTED python '
-        'dev_tools/check_incremental_coverage_annotations.py upstream/master\n'
+        'The annotate command will be removed in a future version.\n'
+        'Get in touch if you still use it: ned@nedbatchelder.com\n'
+        'No data to report.\n'
+        'INTERCEPTED '
+        'python dev_tools/check_incremental_coverage_annotations.py upstream/master\n'
     )
     assert result.err == "Comparing against revision 'upstream/master'.\n"
 
@@ -423,10 +435,13 @@ def test_pytest_and_incremental_coverage_branch_selection(tmpdir_factory):
     assert result.exit_code == 0
     assert result.out == (
         'INTERCEPTED check/pytest '
-        '--actually-quiet --rigetti-integration --cov --cov-report=annotate '
+        '--actually-quiet --rigetti-integration --cov '
         '--cov-config=dev_tools/conf/.coveragerc\n'
-        'INTERCEPTED python '
-        'dev_tools/check_incremental_coverage_annotations.py upstream/master\n'
+        'The annotate command will be removed in a future version.\n'
+        'Get in touch if you still use it: ned@nedbatchelder.com\n'
+        'No data to report.\n'
+        'INTERCEPTED '
+        'python dev_tools/check_incremental_coverage_annotations.py upstream/master\n'
     )
     assert result.err == "Comparing against revision 'upstream/master'.\n"
 
@@ -451,10 +466,13 @@ def test_pytest_and_incremental_coverage_branch_selection(tmpdir_factory):
     assert result.exit_code == 0
     assert result.out == (
         'INTERCEPTED check/pytest '
-        '--actually-quiet --rigetti-integration --cov --cov-report=annotate '
+        '--actually-quiet --rigetti-integration --cov '
         '--cov-config=dev_tools/conf/.coveragerc\n'
-        'INTERCEPTED python '
-        'dev_tools/check_incremental_coverage_annotations.py HEAD\n'
+        'The annotate command will be removed in a future version.\n'
+        'Get in touch if you still use it: ned@nedbatchelder.com\n'
+        'No data to report.\n'
+        'INTERCEPTED '
+        'python dev_tools/check_incremental_coverage_annotations.py HEAD\n'
     )
     assert result.err == "Comparing against revision 'HEAD'.\n"
 
@@ -467,10 +485,13 @@ def test_pytest_and_incremental_coverage_branch_selection(tmpdir_factory):
     assert result.exit_code == 0
     assert result.out == (
         'INTERCEPTED check/pytest '
-        '--actually-quiet --rigetti-integration --cov --cov-report=annotate '
+        '--actually-quiet --rigetti-integration --cov '
         '--cov-config=dev_tools/conf/.coveragerc\n'
-        'INTERCEPTED python '
-        'dev_tools/check_incremental_coverage_annotations.py master\n'
+        'The annotate command will be removed in a future version.\n'
+        'Get in touch if you still use it: ned@nedbatchelder.com\n'
+        'No data to report.\n'
+        'INTERCEPTED '
+        'python dev_tools/check_incremental_coverage_annotations.py master\n'
     )
     assert result.err == "Comparing against revision 'master'.\n"
 
@@ -490,10 +511,13 @@ def test_pytest_and_incremental_coverage_branch_selection(tmpdir_factory):
     assert result.exit_code == 0
     assert result.out.startswith(
         'INTERCEPTED check/pytest '
-        '--actually-quiet --rigetti-integration --cov --cov-report=annotate '
+        '--actually-quiet --rigetti-integration --cov '
         '--cov-config=dev_tools/conf/.coveragerc\n'
-        'INTERCEPTED python '
-        'dev_tools/check_incremental_coverage_annotations.py '
+        'The annotate command will be removed in a future version.\n'
+        'Get in touch if you still use it: ned@nedbatchelder.com\n'
+        'No data to report.\n'
+        'INTERCEPTED '
+        'python dev_tools/check_incremental_coverage_annotations.py '
     )
     assert result.err.startswith("Comparing against revision 'master' (merge base ")