Allow parallel execution of check/pytest scripts in GitHub CI workflows.