@pavoljuhas 2022-02-20
Here is the [current timing](https://github.com/quantumlib/Cirq/runs/5255024345?check_suite_focus=true) of the affected CI jobs compared to the baseline commit; all times are in seconds.  I did not include the *Changed files test*, because number of executed tests can be completely different.  The times are likely subject to the node selection and node load on the GH side, that said we can expect about 5 min shortening of test times (e.g., for Coverage 18m 37s --> 13m 45s).

| CI job | master at 59b72a1ee5ee7c7bec79ad39ef7253f6987bcb16 | PR at fbafc1d66bc75255bbc943a88b60d59260120c3e | delta (s) |
| --- | --- | --- | --- |
| Coverage check | 1117 | 825 | -292 |
| Pytest Ubuntu 3.9 | 587 | 465 | -122 |
| Pytest Windows 3.9 | 874 | 554 | -320 |
| Pytest MacOS | 821 | 392 | -429 |